+++ src/utils/system_utils/src/system_utils.c
@@ -97,13 +97,13 @@ char* ADUC_SystemUtils_MkTemp(char* tmpl)
     // Get a randomish number from real time clock.
     struct timespec ts;
     ADUCPAL_clock_gettime(CLOCK_REALTIME, &ts);
-    unsigned long rnd = ts.tv_nsec ^ (unsigned long)ts.tv_sec;
+    unsigned long rnd = (unsigned long)ts.tv_nsec ^ (unsigned long)ts.tv_sec;
 
     while (*Xs == 'X')
     {
         // '0'..'9' (10), 'A'..'Z' (26), 'a'..'z' (26)
         const unsigned short count = 10 + 26 + 26;
-        const unsigned short idx = (rnd % count);
+        const unsigned short idx = (unsigned short)(rnd % count);
         *Xs = (idx < 10)      ? (char)('0' + idx)
             : (idx < 10 + 26) ? (char)('A' + (idx - 10))
                               : (char)('a' + (idx - (10 + 26)));
